Fixed bug with SubView validation
Originally, "this.valid" was set in an async axios call.
Most of the time, "this.valid" would be set after "update()" finishes,
causing the weird bug of the next clicked on sub having the "this.valid"
of the previous sub.
This issue is fixed by having "get_threads()" call update on success.